Refactoring. Separate factory from resources definition. Allows reuse of resources definitions.
