+++ service/src/main/java/bio/terra/landingzone/library/landingzones/definition/ArmManagers.java
@@ -3,6 +3,7 @@ package bio.terra.landingzone.library.landingzones.definition;
 import com.azure.resourcemanager.AzureResourceManager;
 import com.azure.resourcemanager.applicationinsights.ApplicationInsightsManager;
 import com.azure.resourcemanager.batch.BatchManager;
+import com.azure.resourcemanager.compute.ComputeManager;
 import com.azure.resourcemanager.loganalytics.LogAnalyticsManager;
 import com.azure.resourcemanager.monitor.MonitorManager;
 import com.azure.resourcemanager.postgresql.PostgreSqlManager;
@@ -16,4 +17,5 @@ public record ArmManagers(
     PostgreSqlManager postgreSqlManager,
     LogAnalyticsManager logAnalyticsManager,
     MonitorManager monitorManager,
-    ApplicationInsightsManager applicationInsightsManager) {}
+    ApplicationInsightsManager applicationInsightsManager,
+    ComputeManager computeManager) {}
