+++ service/src/main/java/bio/terra/landingzone/library/landingzones/definition/factories/FakeResourcesFactoryDefinitionV1.java
@@ -1,0 +1,37 @@
+package bio.terra.landingzone.library.landingzones.definition.factories;
+
+import bio.terra.landingzone.library.landingzones.definition.ArmManagers;
+import bio.terra.landingzone.library.landingzones.definition.DefinitionContext;
+import bio.terra.landingzone.library.landingzones.deployment.LandingZoneDeployment;
+import bio.terra.landingzone.library.landingzones.deployment.ResourcePurpose;
+
+// Definition of Fake LZ contains all the resource from the CromwellBase LZ, plus additional
+// resources
+class FakeResourcesFactoryDefinitionV1 extends CromwellBaseResourcesDefinitionV1 {
+
+  public FakeResourcesFactoryDefinitionV1(ArmManagers armManagers) {
+    super(armManagers);
+  }
+
+  @Override
+  public LandingZoneDeployment.DefinitionStages.Deployable definition(
+      DefinitionContext definitionContext) {
+    var base = super.definition(definitionContext);
+
+    /*add resources specific to Fake LZ*/
+
+    var disk =
+        armManagers
+            .computeManager()
+            .disks()
+            .define("FakeLZDisk")
+            .withRegion(definitionContext.resourceGroup().region())
+            .withExistingResourceGroup(definitionContext.resourceGroup())
+            .withData()
+            .withSizeInGB(120)
+            .withTag("purpose", "disk is specific to Fake LZ");
+
+    base.withResourceWithPurpose(disk, ResourcePurpose.WLZ_RESOURCE);
+    return base;
+  }
+}
