+++ service/src/test/java/bio/terra/landingzone/library/landingzones/LandingZoneTestFixture.java
@@ -25,7 +25,8 @@ public class LandingZoneTestFixture {
             TestArmResourcesFactory.createPostgreSqlArmClient(),
             TestArmResourcesFactory.createLogAnalyticsArmClient(),
             TestArmResourcesFactory.createMonitorArmClient(),
-            TestArmResourcesFactory.createApplicationInsightsArmClient());
+            TestArmResourcesFactory.createApplicationInsightsArmClient(),
+            TestArmResourcesFactory.createComputeManager());
     resourceGroup =
         TestArmResourcesFactory.createTestResourceGroup(armManagers.azureResourceManager());
     tokenCredential = AzureIntegrationUtils.getAdminAzureCredentialsOrDie();
