+++ service/src/test/java/bio/terra/landingzone/library/landingzones/TestArmResourcesFactory.java
@@ -6,6 +6,7 @@ import com.azure.core.management.profile.AzureProfile;
 import com.azure.resourcemanager.AzureResourceManager;
 import com.azure.resourcemanager.applicationinsights.ApplicationInsightsManager;
 import com.azure.resourcemanager.batch.BatchManager;
+import com.azure.resourcemanager.compute.ComputeManager;
 import com.azure.resourcemanager.loganalytics.LogAnalyticsManager;
 import com.azure.resourcemanager.monitor.MonitorManager;
 import com.azure.resourcemanager.postgresql.PostgreSqlManager;
@@ -31,7 +32,8 @@ public class TestArmResourcesFactory {
         createPostgreSqlArmClient(),
         createLogAnalyticsArmClient(),
         createMonitorArmClient(),
-        createApplicationInsightsArmClient());
+        createApplicationInsightsArmClient(),
+        createComputeManager());
   }
 
   public static RelayManager createRelayArmClient() {
@@ -66,6 +68,12 @@ public class TestArmResourcesFactory {
 
   public static ApplicationInsightsManager createApplicationInsightsArmClient() {
     return ApplicationInsightsManager.authenticate(
+        AzureIntegrationUtils.getAdminAzureCredentialsOrDie(),
+        AzureIntegrationUtils.TERRA_DEV_AZURE_PROFILE);
+  }
+
+  public static ComputeManager createComputeManager() {
+    return ComputeManager.authenticate(
         AzureIntegrationUtils.getAdminAzureCredentialsOrDie(),
         AzureIntegrationUtils.TERRA_DEV_AZURE_PROFILE);
   }
