+++ service/src/test/java/bio/terra/landingzone/library/landingzones/definition/factories/DeployedLandingZoneDefinitionProviderImplTest.java
@@ -14,7 +14,7 @@ class DeployedLandingZoneDefinitionProviderImplTest {
 
   @BeforeEach
   void setUp() {
-    ArmManagers armManagers = new ArmManagers(null, null, null, null, null, null, null);
+    ArmManagers armManagers = new ArmManagers(null, null, null, null, null, null, null, null);
     provider = new LandingZoneDefinitionProviderImpl(armManagers);
   }
